release

+++ src/installer.ts
@@ -5,8 +5,13 @@ import AdmZip from "adm-zip";
 import * as http from "https"
 import * as util from "util"
 import * as p from "child_process"
+import Store from 'electron-store';
 
-export default function(win: BrowserWindow) {
+
+
+export default function(win: BrowserWindow, store: Store<any>) {
+  
+  let installLocation = store.get('installLocation')
   ipcMain.on('locationSelect', async (e, installLocation: string) => {
     console.log(installLocation)
     const loc = await dialog.showOpenDialog({
@@ -16,66 +21,72 @@ export default function(win: BrowserWindow) {
       defaultPath: installLocation || undefined,
     })
     if (!loc.canceled){
+      store.set('installLocation', loc.filePaths[0])
       e.sender.send('locationSelect', loc.filePaths[0])
     }
   })
   ipcMain.on('isInstalled', (e, gameName: string) => {
-    win.webContents.executeJavaScript('localStorage.getItem("installLocation");', true)
-    .then(result => {
+    
       let installed = false;
-      if (result) installed = fs.existsSync(path.join(result, gameName || ''))
+      if (installLocation) installed = fs.existsSync(path.join(installLocation, gameName || ''))
       e.sender.send('isInstalled', installed)
-    });
   })
 
   ipcMain.on('install', async (e, name: string, url: string)=> {
-    const location = await win.webContents.executeJavaScript('localStorage.getItem("installLocation");', true)
-    const res = http.get(await getDownloadUrl(url), function(response) {
-        var len = parseInt(response.headers['content-length'] || '0', 10);
-        // console.log(response.headers);
-        
-        var cur = 0;
-        var total = len / 1048576; //1048576 - bytes in  1Megabyte
-
-        response.on("data", function(chunk) {
-            cur += chunk.length;
-            e.sender.send('downloadProgress', cur, len)
-            console.log(`Progress: ${cur}/${len}`);
-        });
-
-        res.on("error", function(e){
-            console.log("Error: " + e.message);
-        });
-        const zipPath = path.join(location, 'temp.zip')
-        const stream = fs.createWriteStream(zipPath);
-        response.pipe(stream)
-        stream.on('finish', () => {
-            const zip = new AdmZip(zipPath)
-            zip.extractAllTo(path.join(location, name), true)
-            e.sender.send('installFinished')
-            fs.rmSync(zipPath)
-        })
-    });
-    console.log(util.inspect(res.getHeaders(), false, null, true /* enable colors */))
+    try { 
+        if(!installLocation) throw new Error('Install location not set.')
+        const res = http.get(await getDownloadUrl(url), function(response) {
+            var len = parseInt(response.headers['content-length'] || '0', 10);
+            // console.log(response.headers);
+            
+            var cur = 0;
+            var total = len / 1048576; //1048576 - bytes in  1Megabyte
     
+            response.on("data", function(chunk) {
+                cur += chunk.length;
+                e.sender.send('downloadProgress', cur, len)
+                console.log(`Progress: ${cur}/${len}`);
+            });
+    
+            res.on("error", function(e){
+                console.log("Error: " + e.message);
+            });
+            const zipPath = path.join(installLocation, 'temp.zip')
+            const stream = fs.createWriteStream(zipPath);
+            response.pipe(stream)
+            stream.on('finish', () => {
+                const zip = new AdmZip(zipPath)
+                zip.extractAllTo(path.join(installLocation, name), true)
+                e.sender.send('installFinished')
+                fs.rmSync(zipPath)
+            })
+        });
+        console.log(util.inspect(res.getHeaders(), false, null, true /* enable colors */))
+    }
+    catch (err) {
+        e.sender.send('error', err)
+    }
   })
 
   ipcMain.on('uninstall', async (e, name: string) => {
-    const loc = await win.webContents.executeJavaScript('localStorage.getItem("installLocation");', true)
-    fs.rmSync(path.join(loc, name), {recursive: true, force: true})
+    fs.rmSync(path.join(installLocation, name), {recursive: true, force: true})
     e.sender.send('gameRemoved')
   })
 
   ipcMain.on('play', async (e, name: string) => {
-    const loc = await win.webContents.executeJavaScript('localStorage.getItem("installLocation");', true)
     console.log('play');
-    const executables = recFindByExt(path.join(loc, name), process.platform === 'linux' ? 'x86_64' : 'exe')
+    const executables = recFindByExt(path.join(installLocation, name), process.platform === 'linux' ? 'x86_64' : 'exe')
     win.minimize()
+    let executable = ''
     if(process.platform === 'linux'){
-      p.exec(executables[0]).on('exit', () => {
-        win.maximize()
-      })
+      executable = executables[0]
     }
+    else if(process.platform === 'win32'){
+      executable = executables.find(e => !e.match(/UnityCrashHandler64/)) || ''
+    }
+    p.exec(executable).on('exit', () => {
+      win.show()
+    })
   })
 }
